refactor(tests): expose `advanceRobolectricLooperUntil`
