+++ AnkiDroid/src/test/java/com/ichi2/anki/RobolectricTest.kt
@@ -87,6 +87,9 @@ import org.robolectric.shadows.ShadowLooper
 import org.robolectric.shadows.ShadowMediaPlayer
 import timber.log.Timber
 import kotlin.test.assertNotNull
+import kotlin.time.Duration
+import kotlin.time.Duration.Companion.seconds
+import kotlin.time.TimeSource
 
 open class RobolectricTest :
     AnkiTest,
@@ -271,6 +274,29 @@ open class RobolectricTest :
         // Robolectric needs a manual advance in PAUSED looper mode
         fun advanceRobolectricLooper() {
             Shadows.shadowOf(Looper.getMainLooper()).runToEndOfTasks()
+        }
+
+        /**
+         * Advances the main looper until [condition] holds, failing if [timeout] elapses first.
+         *
+         * @throws IllegalStateException [timeout] has elapsed without [condition] being true.
+         */
+        fun advanceRobolectricLooperUntil(
+            timeout: Duration = 10.seconds,
+            lazyMessage: () -> Any = { "condition not met after $timeout" },
+            condition: () -> Boolean,
+        ) {
+            val start = TimeSource.Monotonic.markNow()
+            while (!condition()) {
+                check(start.elapsedNow() < timeout, lazyMessage)
+                // a real sleep, so background threads finish and post to main
+                Thread.sleep(10)
+                // `advanceRobolectricLooper` only drains tasks already on the main looper,
+                // so it can return while a diff is still in flight (as it's on a different thread).
+                advanceRobolectricLooper()
+            }
+            // flush the work triggered by the condition becoming true (e.g. a layout pass)
+            advanceRobolectricLooper()
         }
 
         @JvmStatic // Using protected members which are not @JvmStatic in the superclass companion is unsupported yet
